+++ packages/backend/src/tests/propertyBased/properties/cardinalityOne.test.ts
@@ -3,7 +3,7 @@ import { DB, schema } from '../../generated';
 import { describe, expect } from 'vitest';
 import { generateArbitraryQuery } from '../arbitraries/arbitraryQuery';
 import { pool, queryEngine } from '../../queryEngine';
-import { executeQuery, getTableRowsByTableName } from './executeQuery';
+import { executeAndWait, getTableRowsByTableName } from './executeAndWait';
 import { CardinalityError } from '../../../query/applyCardinality';
 
 describe('cardinalityOne', async () => {
@@ -24,7 +24,7 @@ describe('cardinalityOne', async () => {
     it.prop([validWhereArbitraryQuery], { verbose: 2 })(
         'Valid where query should return a non-null, non-array, TS object result',
         async (query) => {
-            const queryResult = await executeQuery(queryEngine, query);
+            const queryResult = await executeAndWait(queryEngine, query);
 
             expect(queryResult).toBeTypeOf('object');
 
@@ -38,7 +38,7 @@ describe('cardinalityOne', async () => {
         'Invalid where query should throw error',
         async (query) => {
             try {
-                await executeQuery(queryEngine, query);
+                await executeAndWait(queryEngine, query);
             } catch (error) {
                 expect(error).toBeInstanceOf(CardinalityError);
             }
